♻️ refactor(agent-runtime): split RuntimeExecutors.ts into per-instruction modules
The 4539-line RuntimeExecutors.ts is reduced to a ~40-line assembly factory.
Each instruction executor closes over only `ctx`, so they extract cleanly into
their own files with no behavior change; the public API
(createRuntimeExecutors / RuntimeExecutorContext) is unchanged.

- context.ts: RuntimeExecutorContext interface
- executorHelpers.ts: shared pure helpers (retry budgets, sub-agent/member
  runners, tool discovery, etc.)
- executors/{callLlm,callTool,callToolsBatch,compressContext,execSubAgent,
  finish,humanApprove,resolveTools}.ts: one file per instruction type

type-check clean, RuntimeExecutors.test.ts 125/125 pass.
